+++ sentry/src/test/java/io/sentry/internal/modules/ResourcesModulesLoaderTest.kt
@@ -35,9 +35,9 @@ class ResourcesModulesLoaderTest {
       fixture.getSut(
         content =
           """
-                    com.squareup.okhttp3:okhttp:3.14.9
-                    com.squareup.okio:okio:1.17.2
-                    """
+          com.squareup.okhttp3:okhttp:3.14.9
+          com.squareup.okio:okio:1.17.2
+          """
             .trimIndent()
       )
 
@@ -53,9 +53,9 @@ class ResourcesModulesLoaderTest {
       fixture.getSut(
         content =
           """
-                    com.squareup.okhttp3:okhttp:3.14.9
-                    com.squareup.okio:okio:1.17.2
-                    """
+          com.squareup.okhttp3:okhttp:3.14.9
+          com.squareup.okio:okio:1.17.2
+          """
             .trimIndent()
       )
 
@@ -84,8 +84,8 @@ class ResourcesModulesLoaderTest {
       fixture.getSut(
         content =
           """
-                    com.squareup.okhttp3;3.14.9
-                    """
+          com.squareup.okhttp3;3.14.9
+          """
             .trimIndent()
       )